+++ operatorsDelta.h
@@ -6,8 +6,8 @@
 
 struct DeltaModularityHash : public thrust::unary_function <thrust::tuple<unsigned long long, float>, float> {
 	double* community_weight;
-	float* nodes_weight;
-	float total_weight;
+	unsigned* nodes_weight;
+	double total_weight;
 	float* self_community;
 	unsigned int* communities;
 
@@ -21,15 +21,15 @@ struct DeltaModularityHash : public thrust::unary_function <thrust::tuple<unsign
 		return value;
 	};
 
-	DeltaModularityHash(double* w, float* n, double m, float* s, unsigned int* c) :
+	DeltaModularityHash(double* w, unsigned* n, double m, float* s, unsigned int* c) :
 		community_weight(w), nodes_weight(n), total_weight(m), self_community(s), communities(c) {};
 };
 
 
 struct DeltaModularitySort : public thrust::unary_function <thrust::tuple<unsigned int, unsigned int, float>, float> {
 	double* community_weight;
-	float* nodes_weight;
-	float total_weight;
+	unsigned* nodes_weight;
+	double total_weight;
 	float* self_community;
 	unsigned int* communities;
 
@@ -40,7 +40,7 @@ struct DeltaModularitySort : public thrust::unary_function <thrust::tuple<unsign
 		return value;
 	};
 
-	DeltaModularitySort(double* w, float* n, double m, float* s, unsigned int* c) :
+	DeltaModularitySort(double* w, unsigned* n, double m, float* s, unsigned int* c) :
 		community_weight(w), nodes_weight(n), total_weight(m), self_community(s), communities(c) {};
 };
